Temporary split out core code into seperate directory
