+++ core/core_lib/plugin.py
